Rename utils/backend.py into framework.py for consistency. Fix one more torch import to using `try_import_torch()` in test_nested_spaces.py.
